Yippie!! (Insert TBH creature here)

+++ scripts/cat/cats.py
@@ -2227,8 +2227,7 @@ class Cat:
 
         if (
             (not self.is_ill() and not self.is_injured() and not self.is_disabled())
-            or self.dead
-            or self.status.is_outsider
+            or ((self.dead or self.status.is_outsider) and not self.is_disabled())
         ):
             if os.path.exists(condition_file_path):
                 os.remove(condition_file_path)
